[ExeFramework] fix misnamed new debugger plugins
