--- framework/execution_framework/plugins/org.executionframework.debugger.ui/META-INF/MANIFEST.MF
@@ -1,33 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Ui
-Bundle-SymbolicName: org.executionframework.debugger.ui;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-Activator: org.executionframework.debugger.ui.Activator
-Export-Package: org.executionframework.debugger.ui;uses:="org.gemoc.xdsmlframework.api.core,org.eclipse.core.commands,org.eclipse.jface.viewers",
- org.executionframework.debugger.ui.breakpoints
-Require-Bundle: com.google.guava,
- org.eclipse.xtext.xbase.lib,
- org.eclipse.xtend.lib,
- org.eclipse.xtend.lib.macro,
- org.eclipse.core.commands,
- org.gemoc.xdsmlframework.api,
- org.eclipse.jface,
- org.eclipse.debug.core,
- fr.obeo.dsl.debug.ide,
- org.eclipse.core.runtime;bundle-version="3.12.0",
- org.gemoc.executionframework.engine,
- org.gemoc.commons.eclipse.ui,
- org.eclipse.equinox.registry,
- org.eclipse.ui.workbench,
- org.eclipse.xtext;bundle-version="2.10.0",
- org.executionframework.debugger;bundle-version="1.0.0",
- fr.obeo.dsl.debug.ide.sirius.ui,
- org.gemoc.executionframework.engine.ui,
- fr.inria.diverse.melange.resource,
- fr.obeo.dsl.debug.ide.ui;bundle-version="2.3.0",
- org.eclipse.sirius;bundle-version="4.1.3",
- org.eclipse.debug.ui
-Bundle-ActivationPolicy: lazy
-
