--- framework/execution_framework/plugins/org.executionframework.debugger.ui/plugin.xml
@@ -1,86 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<?eclipse version="3.0"?>
-<plugin>
-    <extension
-          id="org.executionframework.debugger.breakpointMarker"
-          point="org.eclipse.core.resources.markers">
-       <persistent
-             value="true">
-       </persistent>
-       <super
-             type="fr.obeo.dsl.debug.ide.breakpointMarker">
-       </super>
-    </extension>
-    <extension
-          point="org.eclipse.debug.core.breakpoints">
-       <breakpoint
-             class="org.executionframework.debugger.GemocBreakpoint"
-             id="org.executionframework.debugger.breakpoint"
-             markerType="org.executionframework.debugger.ui.org.executionframework.debugger.breakpointMarker"
-             name="Gemoc Breakpoints">
-       </breakpoint>
-    </extension>
-
- <extension
-       point="org.eclipse.debug.ui.detailPaneFactories">
-    <detailFactories
-          class="org.executionframework.debugger.ui.breakpoints.BreakpointDetailPaneFactory"
-          id="org.executionframework.debugger.detailFactories">
-       <enablement>
-          <with
-                variable="selection">
-             <count
-                   value="1">
-             </count>
-             <iterate>
-                <instanceof
-                      value="org.executionframework.debugger.GemocBreakpoint">
-                </instanceof>
-             </iterate>
-          </with>
-       </enablement>
-    </detailFactories>
- </extension>
- <extension
-         point="org.eclipse.ui.bindings">
-      <key
-            commandId="org.executionframework.debugger.ui.openSemantics"
-            contextId="org.eclipse.debug.ui.debugging"
-            schemeId="org.eclipse.ui.defaultAcceleratorConfiguration"
-            sequence="M1+6">
-      </key>
-   </extension>
-
-   <extension
-         point="org.eclipse.ui.commands">
-      <command
-            id="org.executionframework.debugger.ui.openSemantics"
-            name="Open Semantics">
-      </command>
-   </extension>
-   <extension
-         point="org.eclipse.ui.handlers">
-      <handler
-            class="org.executionframework.debugger.ui.OpenSemanticsHandler"
-            commandId="org.executionframework.debugger.ui.openSemantics">
-      </handler>
-   </extension>
-   <extension
-         point="org.eclipse.ui.menus">
-      <menuContribution
-            allPopups="false"
-            locationURI="popup:org.eclipse.debug.ui.DebugView">
-         <command
-               commandId="org.executionframework.debugger.ui.openSemantics"
-               label="Open Semantics">
-         </command>
-      </menuContribution>
-   </extension>
-   <extension
-         point="org.eclipse.sirius.externalJavaAction">
-      <javaActions
-            actionClass="org.executionframework.debugger.ui.breakpoints.GemocToggleBreakpointAction"
-            id="org.executionframework.debugger.ui.GemocToggleBreakpointAction">
-      </javaActions>
-   </extension>
-</plugin>
