--- framework/execution_framework/plugins/org.executionframework.debugger/META-INF/MANIFEST.MF
@@ -1,32 +1,0 @@
-Manifest-Version: 1.0
-Bundle-ManifestVersion: 2
-Bundle-Name: Debugger
-Bundle-SymbolicName: org.executionframework.debugger;singleton:=true
-Bundle-Version: 1.0.0.qualifier
-Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Bundle-Activator: org.executionframework.debugger.Activator
-Require-Bundle: com.google.guava,
- org.eclipse.xtext.xbase.lib,
- org.eclipse.xtend.lib,
- org.eclipse.xtend.lib.macro,
- fr.obeo.dsl.debug.ide,
- org.gemoc.xdsmlframework.api,
- org.eclipse.emf.ecore,
- fr.inria.diverse.trace.commons.model,
- org.eclipse.ui.workbench,
- org.eclipse.swt;bundle-version="3.105.3",
- org.eclipse.jface;bundle-version="3.12.2",
- org.eclipse.debug.ui,
- fr.obeo.dsl.debug;bundle-version="2.3.0",
- org.eclipse.core.runtime;bundle-version="3.12.0",
- org.gemoc.executionframework.engine;bundle-version="2.3.0",
- org.eclipse.emf.transaction;bundle-version="1.9.0",
- org.gemoc.xdsmlframework.commons;bundle-version="2.3.0",
- org.eclipse.xtext;bundle-version="2.10.0",
- org.gemoc.commons.eclipse;bundle-version="2.3.0",
- org.gemoc.commons.eclipse.pde;bundle-version="2.3.0",
- fr.inria.diverse.commons.eclipse;bundle-version="1.0.0",
- fr.inria.diverse.commons.eclipse.messagingsystem.api;bundle-version="1.0.0"
-Export-Package: org.executionframework.debugger
-Bundle-ActivationPolicy: lazy
-
