+++ framework/execution_framework/plugins/org.gemoc.executionframework.debugger/pom.xml
@@ -1,0 +1,65 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+
+  <parent>
+    <relativePath>../..</relativePath>
+    <groupId>org.gemoc.modeldebugging.executionframework</groupId>
+    <artifactId>org.executionframework.debugger</artifactId>
+    <version>2.3.0-SNAPSHOT</version>
+  </parent>
+
+  	<artifactId>org.gemoc.executionframework.engine</artifactId>
+	<version>2.3.0-SNAPSHOT</version>
+  	<packaging>eclipse-plugin</packaging>
+<!--   <build> -->
+<!-- 	  <plugins> -->
+<!-- 	  	<plugin> -->
+<!-- 			<groupId>org.eclipse.tycho</groupId> -->
+<!-- 			<artifactId>tycho-compiler-plugin</artifactId> -->
+<!-- 			<configuration> -->
+<!-- 				<compilerArgument>-warn:+discouraged,forbidden</compilerArgument> -->
+<!-- 			</configuration> -->
+<!-- 		</plugin> -->
+<!-- 	  </plugins>	 -->
+<!--   </build> -->
+
+
+
+	<build>
+		<plugins>
+			<plugin>
+				<groupId>org.eclipse.xtend</groupId>
+				<artifactId>xtend-maven-plugin</artifactId>
+				<version>${xtend.version}</version>
+				<executions>
+					<execution>
+						<goals>
+							<goal>compile</goal>
+						</goals>
+						<configuration>
+							<outputDirectory>xtend-gen</outputDirectory>
+						</configuration>
+					</execution>
+				</executions>
+			</plugin>
+
+			<plugin>
+				<groupId>org.apache.maven.plugins</groupId>
+				<artifactId>maven-clean-plugin</artifactId>
+				<version>2.5</version>
+				<configuration>
+					<filesets>
+						<fileset>
+							<directory>xtend-gen</directory>
+						</fileset>
+					</filesets>
+				</configuration>
+			</plugin>
+		</plugins>
+	</build>
+
+
+
+</project>
