+++ java_execution/java_engine/plugins/org.gemoc.execution.sequential.javaengine.ui/META-INF/MANIFEST.MF
@@ -35,8 +35,8 @@ Require-Bundle: org.eclipse.core.expressions;bundle-version="3.4.400",
  org.eclipse.xtext,
  org.eclipse.xtend.lib,
  fr.inria.diverse.melange.resource,
- org.executionframework.debugger,
- org.executionframework.debugger.ui
+ org.gemoc.executionframework.debugger,
+ org.gemoc.executionframework.debugger.ui
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Bundle-ActivationPolicy: lazy
 Bundle-ClassPath:  .
