+++ java_execution/java_engine/plugins/org.gemoc.execution.sequential.javaengine.ui/src/org/gemoc/execution/sequential/javaengine/ui/commands/GemocToggleBreakpointHandler.java
@@ -24,8 +24,8 @@ import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.ui.ISelectionService;
 import org.eclipse.ui.PlatformUI;
 import org.eclipse.ui.handlers.HandlerUtil;
-import org.executionframework.debugger.GemocBreakpoint;
 import org.gemoc.execution.sequential.javaengine.ui.launcher.Launcher;
+import org.gemoc.executionframework.debugger.GemocBreakpoint;
 
 import fr.inria.diverse.melange.resource.MelangeResource;
 import fr.inria.diverse.trace.commons.model.trace.MSE;
