+++ java_execution/java_engine/plugins/org.gemoc.execution.sequential.javaengine.ui/src/org/gemoc/execution/sequential/javaengine/ui/debug/GenericSequentialModelDebugger.java
@@ -22,9 +22,9 @@ import org.eclipse.emf.ecore.EcorePackage;
 import org.eclipse.emf.ecore.resource.Resource;
 import org.eclipse.xtext.naming.DefaultDeclarativeQualifiedNameProvider;
 import org.eclipse.xtext.naming.QualifiedName;
-import org.executionframework.debugger.AbstractGemocDebugger;
-import org.executionframework.debugger.GemocBreakpoint;
 import org.gemoc.execution.sequential.javaengine.PlainK3ExecutionEngine;
+import org.gemoc.executionframework.debugger.AbstractGemocDebugger;
+import org.gemoc.executionframework.debugger.GemocBreakpoint;
 import org.gemoc.executionframework.engine.core.EngineStoppedException;
 import org.gemoc.xdsmlframework.api.core.IExecutionEngine;
