+++ java_execution/java_engine/plugins/org.gemoc.execution.sequential.javaengine.ui/src/org/gemoc/execution/sequential/javaengine/ui/launcher/Launcher.java
@@ -22,16 +22,16 @@ import org.eclipse.core.runtime.IProgressMonitor;
 import org.eclipse.debug.core.ILaunchConfiguration;
 import org.eclipse.debug.core.ILaunchConfigurationWorkingCopy;
 import org.eclipse.emf.ecore.EObject;
-import org.executionframework.debugger.AbstractGemocDebugger;
-import org.executionframework.debugger.AnnotationMutableFieldExtractor;
-import org.executionframework.debugger.IMutableFieldExtractor;
-import org.executionframework.debugger.IntrospectiveMutableFieldExtractor;
+import org.gemoc.executionframework.debugger.AnnotationMutableFieldExtractor;
+import org.gemoc.executionframework.debugger.IntrospectiveMutableFieldExtractor;
 import org.gemoc.commons.eclipse.ui.ViewHelper;
 import org.gemoc.execution.sequential.javaengine.PlainK3ExecutionEngine;
 import org.gemoc.execution.sequential.javaengine.SequentialModelExecutionContext;
 import org.gemoc.execution.sequential.javaengine.ui.Activator;
 import org.gemoc.execution.sequential.javaengine.ui.debug.GenericSequentialModelDebugger;
 import org.gemoc.execution.sequential.javaengine.ui.debug.OmniscientGenericSequentialModelDebugger;
+import org.gemoc.executionframework.debugger.AbstractGemocDebugger;
+import org.gemoc.executionframework.debugger.IMutableFieldExtractor;
 import org.gemoc.executionframework.engine.commons.EngineContextException;
 import org.gemoc.executionframework.engine.commons.ModelExecutionContext;
 import org.gemoc.executionframework.engine.ui.commons.RunConfiguration;
