+++ trace/manager/plugins/org.gemoc.sequential_addons.multidimensional.timeline/src/org/gemoc/sequential_addons/multidimensional/timeline/views/MultidimensionalTimelineViewPart.java
@@ -54,7 +54,7 @@ import org.eclipse.swt.widgets.Menu;
 import org.eclipse.swt.widgets.MenuItem;
 import org.eclipse.swt.widgets.Shell;
 import org.eclipse.ui.IActionBars;
-import org.executionframework.debugger.AbstractGemocDebugger;
+import org.gemoc.executionframework.debugger.AbstractGemocDebugger;
 import org.gemoc.executionframework.engine.ui.launcher.AbstractGemocLauncher;
 import org.gemoc.executionframework.ui.views.engine.EngineSelectionDependentViewPart;
 import org.gemoc.executionframework.ui.views.engine.actions.AbstractEngineAction;
